Move workflow code to _internal
